requirements.txt
streamlit
agno
groq
python-dotenv